test: move blocking io detector to support
