+++ backend/tests/support/detectors/__init__.py
@@ -1,0 +1,1 @@
+"""Runtime and static detectors used by tests."""
